refactor: moved new data modules to a the pages folder
